refactor(pack): refactor xdsl ip
